backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary   # später für PostgreSQL-Anbindung
